Fix: Supersampling, temporal_score_stddev hardcoded
Super sampling was given as `supersampling` previously, leading to the
parameter being ignored by the backend `super_sampling` is correct.

The Python implementation has two different values for
`temporal_score_stdev` when used on S3 and S2 inputs.
The parameter has been removed and the different values,
10 for S3 and 20 for S2 hardcoded.
The parameters could be added back in later (separately).